requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
torch>=2.0
biopython>=1.83
tqdm>=4.66
pyyaml>=6.0
einops>=0.7
